+++ src/app/components/shared/model-list/model-list.component.scss
@@ -1,6 +1,6 @@
 @use "bootstrap" as *;
 
-.region-map {
+.geographic-map {
   display: block;
   width: 100%;
   height: 40rem;
